Move to :core:database module.
