+++ Compose-Flight-Search-App/core/database/src/main/kotlin/com/example/flightsearchapp/database/model/SuggestionAirport.kt
@@ -1,4 +1,4 @@
-package com.example.flightsearchapp.ui.model
+package com.example.flightsearchapp.database.model
 
 data class SuggestionAirport(
     val id: Long,
