+++ Compose-Flight-Search-App/core/ui/src/main/kotlin/com/example/flightsearchapp/ui/flight/FlightsListScreen.kt
@@ -5,10 +5,10 @@ import androidx.compose.foundation.lazy.items
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.flightsearchapp.ui.SearchScreenUiState
-import com.example.flightsearchapp.ui.ShowFlightUiState
+import com.example.flightsearchapp.core.designsystem.theme.FlightSearchAppTheme
 import com.example.flightsearchapp.ui.model.Flight
-import com.example.flightsearchapp.ui.theme.FlightSearchAppTheme
+import com.example.flightsearchapp.ui.uistate.SearchScreenUiState
+import com.example.flightsearchapp.ui.uistate.ShowFlightUiState
 
 @Composable
 fun FlightsListScreen(
